+++ src/RunBook/Task.ts
@@ -26,12 +26,37 @@ export class Task {
     const a = this.task.operation.split(".");
     const op = winOperations.filter(this.task.operation);
 
+    // app.exportMany is slightly different.
+    // And should be handled separately :(
+    if ((this.task.operation as string) == "app.exportMany") {
+      // if "source" is provided then call the export()
+      // method of each object from the source task
+      // no need to call the "filter" again. Will have to
+      // build the filter manually and this can lead to some issues
+      if (!this.task.filter && this.task.source) {
+        return await Promise.all(
+          this.objectsData.data.map((o) => {
+            return o["export"]({ ...this.task.details });
+          })
+        );
+      }
+
+      // if "filter" is provided then just pass it as it is
+      return (await this.instance.apps
+        .exportMany({ filter: this.task.filter, ...this.task.details })
+        .catch((e) => {
+          e.message = `REST communication error! ${e.message}`;
+          e.stack = "";
+          throw e;
+        })) as any;
+    }
+
     // some tasks have to be "renamed" to match the corresponding method
     // for example:
     //    virtualProxy -> virtualProxies
     //    customProperty -> customProperties
     if (a[0].substring(a[0].length - 1) == "y" && op.isPlural)
-      a[0] = `${a[0].substring(0, a[0].length - 1)}ies`;
+      a[0] = `${a[0].substring(0, a[0].length - 1)}ie`;
 
     // if the task do NOT require initial data (filter or source)
     // for example about.*, app.import etc
@@ -39,7 +64,8 @@ export class Task {
       if (
         winOperations.nonSourceOperationsPlural.indexOf(this.task.operation) ==
         -1
-      )
+      ) {
+        if (a[0].substring(a[0].length - 2) == "ie") a[0] = `${a[0]}s`;
         return await this.instance[a[0]]
           [a[1]](this.task.details || {}, this.task.options)
           .catch((e) => {
@@ -47,6 +73,7 @@ export class Task {
             e.stack = "";
             throw e;
           });
+      }
 
       return await this.instance[`${a[0]}s`]
         [a[1]](this.task.details || {}, this.task.options)
@@ -58,11 +85,11 @@ export class Task {
     }
 
     // if the task require initial data
-    if (a[1] == "get") return this.objectsData.data; //.map((d) => d.details);
+    if (a[1] == "get" || a[1] == "getAll") return this.objectsData.data; //.map((d) => d.details);
 
-    // check if the previous data is an array
-    // and if its not then call the method directly
     if (!Array.isArray(this.objectsData.data))
+      // check if the previous data is an array
+      // and if its not then call the method directly
       return await this.objectsData.data[a[1]](
         this.task.details || {},
         this.task.options
@@ -72,7 +99,6 @@ export class Task {
     // loop though all elements and call the method in each element
     return await Promise.all(
       this.objectsData.data.map(async (obj) => {
-        let a1 = 1;
         return obj[a[1]](this.task.details, this.task.options);
       })
     );
@@ -97,22 +123,27 @@ export class Task {
 
     // if no objects are found throw an error
     if (this.objectsData.data.length == 0) {
-      if (!this.task.config)
+      if (!this.task.options)
         throw new CustomError(1004, this.task.name, { arg1: this.task.name });
 
-      if (this.task.config.allowZero == false)
+      if (this.task.options.allowZero == false)
         throw new CustomError(1007, this.task.name, { arg1: this.task.name });
     }
 
     // by default task will throw an error if multiple objects are returned and config is missing
-    if (this.objectsData.data.length > 1 && !this.task.config)
+    if (
+      this.objectsData.data.length > 1 &&
+      !this.task.options &&
+      (!this.task.operation.indexOf(".get") ||
+        !this.task.operation.indexOf(".getAll"))
+    )
       throw new CustomError(1005, this.task.name, { arg1: this.task.name });
 
     // multiple objects are returned and config.multiple is set to false
     if (
       this.objectsData.data.length > 1 &&
-      this.task.config &&
-      !this.task.config.multiple
+      this.task.options &&
+      !this.task.options.multiple
     )
       throw new CustomError(1006, this.task.name, { arg1: this.task.name });
   }