+++ src/util/CustomError.ts
@@ -22,9 +22,9 @@ export class CustomError extends Error {
     1002: `Initial checks: Task names duplications. Duplicated names: %{arg1}`,
     1003: `No "source" or "filter" is provided for task "%{arg1}"`,
     1004: `Task "%{arg1}" aborted. 0 objects will be affected`,
-    1005: `Task "%{arg1}" will be performed on multiple objects. Specify "config.multiple = true" to override`,
-    1006: `Task "%{arg1}" will be performed on multiple objects. "config.multiple" is set to "false"`,
-    1007: `Task "%{arg1}" aborted. 0 objects will be affected and "config.allowZero" is set to "false"`,
+    1005: `Task "%{arg1}" will be performed on multiple objects. Specify "options.multiple = true" to override`,
+    1006: `Task "%{arg1}" will be performed on multiple objects. "options.multiple" is set to "false"`,
+    1007: `Task "%{arg1}" aborted. 0 objects will be affected and "options.allowZero" is set to "false"`,
     1008: `Task "%{arg1}" aborted. Task "filter" is provided but the filter is empty`,
     1009: `Task "%{arg1}" aborted. Task "source" is provided but the filter is empty"`,
     1010: `Task "%{arg1}" aborted. Task "operation" is required"`,
@@ -32,6 +32,7 @@ export class CustomError extends Error {
     1012: `Task "%{arg1}" aborted. "operation" property is missing`,
     1013: `Initial checks: Non existing operations found: %{arg1}`,
     1014: `Initial checks: Source/filter for tasks is required: %{arg1}`,
+    1015: `Initial checks: Wrong custom properties names: %{arg1}`,
   };
 
   constructor(code: number, taskName: string, params?: IErrorParams) {